split key mask stuff methods out
